Add compiler type-checking of format strings via the GNU "format"
attribute, annotate the format-taking functions, and fix the mismatches
this exposed.

Changes:
- Add "attr.h" with an ATTR_FORMAT wrapper
- Annotate strdupf, strcatf, and strftimet
- Annotate the log_* functions and the _log_aux helper
- Annotate _path_set_err and remunge's output_msg
- Fix uint32_t and size_t specifier mismatches
- Fix sig_atomic_t specifier mismatches on FreeBSD
- Suppress -Wformat-nonliteral for strftimet's format forwarding

Tested:
- AlmaLinux 10.2 (x86-64-v2), 9.8, 8.10
- Arch Linux
- CentOS Stream 9, 7.9.2009
- Debian sid, 13.6, 12.15, 11.11, 10.13, 9.13, 8.11, 7.11, 6.0.10, 5, 4
- Fedora 44, 43, 42
- FreeBSD 15.1, 15.0, 14.4
- macOS 26.4 (25E246)
- NetBSD 10.1, 9.4
- OpenBSD 7.9, 7.8
- openSUSE Leap 16.0, 15.6
- Ubuntu 26.04, 24.04.4, 22.04.5, 20.04.6, 18.04.6, 16.04.7, 14.04.6

+++ src/munged/job.c
@@ -87,8 +87,8 @@ job_accept (conf_t conf, work_p workers)
 
     while (!got_terminate) {
         if (got_reconfig) {
-            log_msg (LOG_NOTICE, "Processing signal %d (%s)",
-                    got_reconfig, strsignal (got_reconfig));
+            log_msg (LOG_NOTICE, "Processing signal %ld (%s)",
+                    (long) got_reconfig, strsignal (got_reconfig));
             got_reconfig = 0;
             gids_update (conf->gids);
         }
@@ -174,8 +174,8 @@ job_accept (conf_t conf, work_p workers)
             log_msg (LOG_WARNING, "Failed to queue client request");
         }
     }
-    log_msg (LOG_NOTICE, "Exiting on signal %d (%s)",
-            got_terminate, strsignal (got_terminate));
+    log_msg (LOG_NOTICE, "Exiting on signal %ld (%s)",
+            (long) got_terminate, strsignal (got_terminate));
 }
